+++ WolvenKit.App/ViewModels/Dialogs/ShowChecklistDialogueViewModel.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Linq;
 using CommunityToolkit.Mvvm.ComponentModel;
 using DynamicData;
+using WolvenKit.App.Interaction.Options;
 using WolvenKit.App.Models.ProjectManagement.Project;
 using WolvenKit.Interfaces.Extensions;
