+++ WolvenKit.App/ViewModels/Shell/AppViewModel.ComplexFiles.cs
@@ -8,6 +8,7 @@ using CommunityToolkit.Mvvm.Input;
 using Octokit;
 using WolvenKit.App.Helpers;
 using WolvenKit.App.Interaction;
+using WolvenKit.App.Interaction.Options;
 using WolvenKit.App.Models.ProjectManagement.Project;
 using WolvenKit.App.ViewModels.Dialogs;
 using WolvenKit.App.ViewModels.Tools;
