+++ WolvenKit/Views/Dialogs/Windows/ShowDictionaryForCopyDialogView.xaml.cs
@@ -6,6 +6,7 @@ using System.Windows.Data;
 using System.Windows.Input;
 using ReactiveUI;
 using WolvenKit.App.Helpers;
+using WolvenKit.App.Interaction.Options;
 using WolvenKit.App.ViewModels.Dialogs;
 
 namespace WolvenKit.Views.Dialogs.Windows;
