+++ WolvenKit/Views/Documents/RedDocumentViewMenuBar.xaml.cs
@@ -15,6 +15,7 @@ using Splat;
 using WolvenKit.App;
 using WolvenKit.App.Helpers;
 using WolvenKit.App.Interaction;
+using WolvenKit.App.Interaction.Options;
 using WolvenKit.App.Models.ProjectManagement.Project;
 using WolvenKit.App.Services;
 using WolvenKit.App.ViewModels.Dialogs;
